django/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
pipenv==2022.4.8
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
whitenoise==6.0.0
